Memory leak fixes for inputhost
Summary:
The stress test was successful and the latency was good as well
after the change to use channel based close.
This patch simply re-reverts all the leak fixes we already fixed.
I will send the CR for the channel based close separately.

This patch gets all the leak fixes by re-reverting the
commit.
One Unit test needed to be fixed to get the appropriate
metrics reporter to make sure it doesn't crash.

Test Plan: UTs

Reviewers: weihan

Reviewed By: weihan