Move to idiomatic golang's channel based close
Summary:
This makes sure we don't have all callbacks competing
for the path mutex.
With the callback based close, there were a bunch of cbs competing
with the path mutex and the latencies started getting worse.

This patch just moves to a more idiomatic channel based close
which makes sure we unload properly and also not make the latency
worse.

Test Plan: make

Reviewers: weihan

Reviewed By: weihan