Move the flushticker within the replica connection
time.Ticker is not buffered as well, so the solution of just
moving the ticker to exthost will not work for all replicas.

We need to have the ticker on the exthost and then control the flush
from there by sending a notification to each replica. But before
we implement something complex there let's run with the reduced
flush threshold for now and get the numbers.