Add graceful drain of extents on inputhost
This patch simply implements the graceful drain on inputhost.
Whenever an inputhost gets a graceful drain request, it
does the following:
(1) Marks the extent as draining and stops the writepump and waits
    for the write pump to drain (for a timeout period)
(2) On the read pump if we have drained all the messages, then we
    can safely close the extent and go away.

A couple of caveats:
1. Note: close is idempotent and so we should be able to force close
   an extent anytime during the drain process.
2. We just need the RLock() during read pump aggregation and since it
   is not contended the profiling of inputhost didn't reveal it to be
   at top.

Also add a unit test for the same.
I will be adding more unit tests to test the failure paths as well