Add the ability to handle signals
With this patch, we can have handler routines for all kinds of
signals as needed.

For example, if we want to handle SIGTERM then just call
common.HandleSignal() with the appropriate arguments and the handler
function as well, which can be used to perform any activity like say
calling an upgrade endpoint, etc.